fix: split session subscription modules under 500-line cap (#4117)
Re-split consolidated AppEventHandlerScope session listeners into
per-family libraries and extract train-order helpers so
repo.app_event_handler_part_size and main-line budget pass again.
